+++ CHANGELOG.md
@@ -6,6 +6,7 @@ under development
 - Add `Skipper`.
 - Add `PathSkipper`.
 - Add `IRouter.Any`: registers a new request handler function that matches any HTTP methods.
+- Add `Validatable`: if a value is validatable, `Context.Decode` will validate the value.
 
 v1.11.0 April 15, 2020
 ----------------------